Add files via upload
YouTube dataset analysis using pandas, numpy and seaborn. In this 
analysis I will be using TOP5000 youtube channels gotten from 
Kaggle. the objective of this analysis are as follows...

1. using the describe function to convert the data statistics from 
exponential to decimal.
2. replace -- to NaN looking at the "Subscribers" columns you 
will find "--". In this task I want to replace every "--" with 
"NaN"
3. Clean the "Rank Column"(I removed the following in the 
rank column "st", "nd", "rd", "th" and ",") & convert the 
datatypes from object to integer
4. #Convert the datatype of the "Video Uploads & Subscribers" 
columns from object to int.
5. #Clean the "Grade" column by giving each unique grades 
numerical attribute & convert datatype to int
6. The average view for each channel
7. #TOP10 channels with max number of video uploads
8. TOP10 channels with max number of views
9. Find the corr matrix(From my correlation I found out that 
Subcribers and Video views are highly correlated unlike other 
columns.
10. Which grades has the max number of video upload(From my bar
plot it can be seen that the A++ which is 10 has the highest 
vidoe uploads with A+ which is 8 coming a distance 2nd)
11. which grade has the highest avaerage views(Unlike the case 
of video uploads the A+ grade has the highest average views 
followed by A grade at the 2nd position as shown in the barplot)

12. Which grade has the highest number of subscribers(From our 
plot A++ grade channel has the highest number of subscribers 
while the B grade channels have the lowest number of 
subscribers as shown in the visualization).

13. which grade has the highest video views(A++ Grade has the 
highest video views with the B Grade having the lowest video
views)

I checked for missing values, I discovered the "Video uploads" 
has 6 missing values while "subscribers" column recorded 387 
missing values meanwhile the rest has no missing values and the 
percentage are 0.12 for Video "Uploads" and 7.74 for "Subscribers"